+++ src/Homes/Card.jsx
@@ -1,12 +1,21 @@
 import React from 'react';
 import styled from 'styled-components';
-import { Img, Price, Reviews, Rating, Star, VerticalContainer as Container } from '../lib/Card';
-import star from '../imgs/star.svg';
+import { Img, Price, Reviews, Rating, Star, VerticalContainer as Container } from '../UI/Card';
+import star from '../UI/star.svg';
 
-const Name = styled.span`
+const Area = styled.span`
+  font-size: 12px;
+  font-weight: light;
+
+  @media (min-width: 768px) {
+    font-size: 15px;
+  }
+`;
+
+const Title = styled.span`
   font-size: 13px;
-  font-weight: 200;
-  margin-bottom: 5px;
+  font-weight: bold;
+  margin-bottom: 2px;
   margin-top: 7px;
 
   @media (min-width: 768px) {
@@ -14,22 +23,27 @@ const Name = styled.span`
   }
 `;
 
+const PushedReviews = Reviews.extend`
+  margin-top: 6px;
+`;
+
 function Card(props) {
   return (
     <Container href={props.href} title={props.alt}>
       <Img src={props.img} alt={props.alt} title={props.alt} />
-      <Name>
+      <Title>
         <Price>{props.price} </Price>
         {props.title}
-      </Name>
-      <Reviews>
+      </Title>
+      <Area>{props.area}</Area>
+      <PushedReviews>
         <Star src={star} />
         <Star src={star} />
         <Star src={star} />
         <Star src={star} />
         <Star src={star} />
-        <Rating>{props.numOfReviews}</Rating>
-      </Reviews>
+        <Rating>{props.rating}</Rating>
+      </PushedReviews>
     </Container>
   );
 }
