+++ src/Landing/Featured/featured-6.png
